requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==45.0.5
distlib==0.4.0
dlib-bin==19.24.6
ecdsa==0.19.1
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httptools==0.6.4
identify==2.6.15
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
nodeenv==1.9.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pandas==2.0.3
passlib==1.7.4
pillow==11.3.0
platformdirs==4.4.0
pre_commit==4.3.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
reverse_geocoder==1.5.1
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.47.2
stripe==12.5.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1
strawberry-graphql[fastapi]==0.157.0
grpcio==1.76.0
grpcio-tools==1.76.0
psycopg2-binary==2.9.10
datafusion==41.0.0
